04.19 Additional Minor Update
A huge mistake was made when updating the editMember() method. It deleted the map after altering the member's name, without checking if any member records remained. It is fixed now. The list now removes the member, and checks if the character is empty. If the Jaeum is empty, it too will be removed.